Fixed Save/Load Model: lấy dữ liệu về Download của thiết bị điện thoại ở thư folder XHeroScan [Lưu ý: một số điện thoại sẽ không có quyền lưu trực tiếp: cần fix thêm]
