+++ Assets/Prefabs/ErrorPanel.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: fbba4f0e5d996d34e911e0d373c8d376
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
