+++ Assets/Prefabs/SuccessPanel.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: b16a5d162b182be4a8c44f0992629d2c
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
