+++ UserSettings/AndroidLogcatSettings.asset
@@ -1,13 +1,146 @@
 {
-    "m_SelectedDeviceId": "",
+    "m_SelectedDeviceId": "RFCT90BE7FX",
     "m_SelectedProcess": {
-        "deviceId": "",
-        "name": "",
-        "processId": 0,
+        "deviceId": "RFCT90BE7FX",
+        "name": "com.XHeroZone.XHeroScan",
+        "processId": 3226,
         "exited": false
     },
     "m_SelectedPriority": 0,
-    "m_KnownProcessesForSerialization": [],
+    "m_KnownProcessesForSerialization": [
+        {
+            "deviceId": "RFCT90BE7FX",
+            "name": "com.XHeroZone.XHeroScan",
+            "processId": 31551,
+            "exited": false
+        },
+        {
+            "deviceId": "RFCT90BE7FX",
+            "name": "com.XHeroZone.XHeroScan",
+            "processId": 32718,
+            "exited": false
+        },
+        {
+            "deviceId": "RFCT90BE7FX",
+            "name": "com.XHeroZone.XHeroScan",
+            "processId": 1633,
+            "exited": false
+        },
+        {
+            "deviceId": "RFCT90BE7FX",
+            "name": "com.XHeroZone.XHeroScan",
+            "processId": 2884,
+            "exited": false
+        },
+        {
+            "deviceId": "RFCT90BE7FX",
+            "name": "com.XHeroZone.XHeroScan",
+            "processId": 4151,
+            "exited": false
+        },
+        {
+            "deviceId": "RFCT90BE7FX",
+            "name": "com.XHeroZone.XHeroScan",
+            "processId": 5874,
+            "exited": false
+        },
+        {
+            "deviceId": "RFCT90BE7FX",
+            "name": "com.XHeroZone.XHeroScan",
+            "processId": 6382,
+            "exited": false
+        },
+        {
+            "deviceId": "RFCT90BE7FX",
+            "name": "com.XHeroZone.XHeroScan",
+            "processId": 7136,
+            "exited": false
+        },
+        {
+            "deviceId": "RFCT90BE7FX",
+            "name": "com.XHeroZone.XHeroScan",
+            "processId": 21354,
+            "exited": false
+        },
+        {
+            "deviceId": "RFCT90BE7FX",
+            "name": "com.XHeroZone.XHeroScan",
+            "processId": 22290,
+            "exited": false
+        },
+        {
+            "deviceId": "RFCT90BE7FX",
+            "name": "com.XHeroZone.XHeroScan",
+            "processId": 22899,
+            "exited": false
+        },
+        {
+            "deviceId": "RFCT90BE7FX",
+            "name": "com.XHeroZone.XHeroScan",
+            "processId": 23640,
+            "exited": false
+        },
+        {
+            "deviceId": "RFCT90BE7FX",
+            "name": "com.XHeroZone.XHeroScan",
+            "processId": 24339,
+            "exited": false
+        },
+        {
+            "deviceId": "RFCT90BE7FX",
+            "name": "com.XHeroZone.XHeroScan",
+            "processId": 25882,
+            "exited": false
+        },
+        {
+            "deviceId": "RFCT90BE7FX",
+            "name": "com.XHeroZone.XHeroScan",
+            "processId": 26523,
+            "exited": false
+        },
+        {
+            "deviceId": "RFCT90BE7FX",
+            "name": "com.XHeroZone.XHeroScan",
+            "processId": 27231,
+            "exited": false
+        },
+        {
+            "deviceId": "RFCT90BE7FX",
+            "name": "com.XHeroZone.XHeroScan",
+            "processId": 29019,
+            "exited": false
+        },
+        {
+            "deviceId": "RFCT90BE7FX",
+            "name": "com.XHeroZone.XHeroScan",
+            "processId": 29532,
+            "exited": false
+        },
+        {
+            "deviceId": "RFCT90BE7FX",
+            "name": "com.XHeroZone.XHeroScan",
+            "processId": 31834,
+            "exited": false
+        },
+        {
+            "deviceId": "RFCT90BE7FX",
+            "name": "com.XHeroZone.XHeroScan",
+            "processId": 493,
+            "exited": false
+        },
+        {
+            "deviceId": "RFCT90BE7FX",
+            "name": "com.XHeroZone.XHeroScan",
+            "processId": 2438,
+            "exited": false
+        },
+        {
+            "deviceId": "RFCT90BE7FX",
+            "name": "com.XHeroZone.XHeroScan",
+            "processId": 3226,
+            "exited": false
+        }
+    ],
     "m_Tags": {
         "m_Entries": [
             {
