+++ UserSettings/EditorUserSettings.asset
@@ -24,19 +24,19 @@ EditorUserSettings:
       value: 0655075653565d5e5e0d5e2312705c4445161b7b7c7c2063757c1862e4b66261
       flags: 0
     RecentlyUsedSceneGuid-4:
-      value: 5a5757560101590a5d0c0e24427b5d44434e4c7a7b7a23677f2b4565b7b5353a
+      value: 5550555e07575b0c0c585a2111730944434f4b7d7b2c7f617b7a1c64e3e3316f
       flags: 0
     RecentlyUsedSceneGuid-5:
       value: 0005570304565a5a585d557044200b4415154e727a2c7166792f4d65e6b56d3d
       flags: 0
     RecentlyUsedSceneGuid-6:
-      value: 5108510055025c0d590a597042705b444f4e407c7b29246974781863bab76c68
+      value: 5a5757560101590a5d0c0e24427b5d44434e4c7a7b7a23677f2b4565b7b5353a
       flags: 0
     RecentlyUsedSceneGuid-7:
-      value: 5550555e07575b0c0c585a2111730944434f4b7d7b2c7f617b7a1c64e3e3316f
+      value: 0250525753035b020e5f5e2048265d44154e4a7e7b2c7234287e1c6bb4e16c3c
       flags: 0
     RecentlyUsedSceneGuid-8:
-      value: 0250525753035b020e5f5e2048265d44154e4a7e7b2c7234287e1c6bb4e16c3c
+      value: 5108510055025c0d590a597042705b444f4e407c7b29246974781863bab76c68
       flags: 0
     vcSharedLogLevel:
       value: 0d5e400f0650
